requirements.txt
asgiref==3.4.1
Django==4.0.1
django-environ==0.8.1
gunicorn==20.1.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
